simplify dummy function branching
