--- posthog/clickhouse/materialized_columns/__init__.py
@@ -1,10 +1,0 @@
-from posthog.models.property import TableWithProperties
-from posthog.settings import EE_AVAILABLE
-
-ColumnName = str
-TablesWithMaterializedColumns = TableWithProperties
-
-if EE_AVAILABLE:
-    from ee.clickhouse.materialized_columns.columns import get_materialized_columns
-else:
-    from .column import get_materialized_columns
